bot/requirements.txt
flask==2.3.6
google-cloud-dialogflow==2.12.0
streamlit==1.30.0
requests==2.31.0
python-dotenv==1.0.0
# Optional for Telegram polling
python-telegram-bot==20.5
